# backend/internal/extension/schema/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Amitia Extension Runtime API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/extensions/kernel/extension": {
      "get": {
        "description": "Get kernel extension",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Get kernel extension",
        "tags": [
          "Kernel"
        ]
      }
    },
    "/extensions/kernel/extensions": {
      "get": {
        "description": "List kernel extensions",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "List kernel extensions",
        "tags": [
          "Kernel"
        ]
      }
    },
...
    "/extensions/packages": {
      "get": {
        "responses": {
          "200": {
            "description": "Installed extension package list"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "List installed extension packages"
      }
    },
...
    "/extensions/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Uninstall result"
          },
          "409": {
            "description": "Dependency in use"
          }
        },
        "summary": "Uninstall a local Skill while preserving audits and runs"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/api"
    }
  ]
}